[src/main.cpp] Fix turbo getting reset
When turbo value was not manually requested, setting the sane CPU
values would reset the value of Turbo back to the system default,
which is not what pstate-frequency claims to do in those situations.

This patch fixes a small amount of the code logic where the program
would only check that a value has not been requested only steps before
deciding an assignment value. This patch checks for assignment values
before setting sane defaults, and then applys the system current
configuration if in the event of no requested value for a certain CPU
feature like Turbo.