[src/psfreq_strings.c] Do not throw an error in strtol
strtol will return a 0 or error, but the user could also specify 0 as
the string that they wish to convert into a number. Do not throw at
error during converion, and instead check for a valid result later down
the line in each specific instance where a 0 would be unacceptable.