[config.mk] Explicitly set to -std=gnu89
Do not allow the standard to be changed anymore, as the project now
expects gnu89 as the C standard. This should not limit compatability, as
the Linux kernel expects C90 support to compile, and if you can't run
the kernel, you can't run this. If there is a machine out there, running
the intel_pstate driver, but lacking support for C90, please.

Let me know.