[src/psfreq_cpu.{c,h}] Drop support for acpi-cpufreq
For the time being, drop support for systems that do not use
intel_pstate. intel_pstate has been supported since kernel 3.9, if a
user is not using it still, then there are many other tools that they
can use to do the same job that pstate-frequency attempts to fill for
intel_pstate users.