[config.mk] Default udev to true
By default, include the udev file in standard builds. The udev file should be
stable enough that it can be included without too much trouble on most
systems. It also performs similarly to the pstate-frequency.service
systemd service in that on boot it can set the frequencies of the CPU
on first start.