[systemd service] Wait until after sysinit.target to run pstate-frequency
pstate-frequency experiences issues on systems with SSD boot drives, in that the system starts up so fast that pstate-frequency
attempts to run before the system meets sane conditions, particularly in the mounting and presence of the sys and proc
directories on the root.

This attempts to have the pstate-frequency.service wait until after the sysinit target is reached, which generally means
that the system has finished initializing. At this point, it should be safe to call pstate-frequency