[Makefile] Prompt with makefile options by default
By default, invoking just make on the command line will now print out the various arguments that can be passed to make.
Any other targets which make exist in the makefile should be thought of as internal, and not be directly used.

Hopefully this will clean up confusion about using the makefile to build the program.