Work around systemd service issues by adding sleep calls
On systems with fast SSDs the systemd services may fail because certain parts of the sysfs
or devfs are not available at the time of boot. Work around this issue with an ugly hack
by waiting for about 5 seconds before calling the respective services into action.

Modify the makefile for more clear install sections and remove the calls to sed as paths
are resolved when using the shell in a systemd ExecStart call.