Prepare for power plan configurations
Power plans will be specified in a simple configuration file that can be sourced
by the program

Because the script will be sourcing as the root user, there is the potential for damage if something
goes wrong. It is up to you, the user, to make sure that there is no potentially malicious code
that could acidentally be sourced should you select a given power plan.