Update to new branch structure
To better seperate sections of work the project structure is changing slightly.
The 'master' branch will now be rolling, as the dev branch once was. It has effectively
replaced the 'dev' branch and will be the current, bleeding builds of pstate-frequency.

Stable releases will be handled by tags. Each stable release will be tagged. For AUR related builds,
the stable build will first be tagged and pushed, and AFTER the SHA256 sums will be updated for stable
PKGBUILDS. This is not a change, and is how things are always done. If you are expecting to build from the
in tree PKGBUILDs instead of the ones on the AUR, this is a friendly reminder.