Change how plans decide on a CPU governor
Previously, plans would only be able to set a single CPU governor, regardless of CPU driver.
The builds were shipped with governors that would work for acpi-cpufreq systems.

Except that, wait, this is pstate-frequency, not acpi-cpufreq-frequency. The shipped build did not
default to settings that would work with intel_pstate, which is a big problem.

This patch fixes this issue, by defining two seperate variables, one for the CPU governor if the
system is running on acpi-cpufreq, and the other if the system is running on intel_pstate

Plans that do not update to take advtantage of this newer syntax may not work as expected