Some performance optimizations
Reduce additional work by avoiding setters that just use the "current"
value of things if they are not requested.

Before, pstate-frequency would for example, apply -min with the current
system value even if the user did not request a -min on the command
line. This was an extra file read and an extra file write which largely
did nothing as the kernel would have already smartly adjusted the values
in the sysfs. Maybe this would have had an effect on an older or buggy
kernel, but now that we are around v6, the kernel sysfs for CPU stuff is
pretty robust.

This removes a bunch of the useless writes (as they are largely ignored
anyway on modern hardware pstate systems)