amd-pstate touches cpuinfo_max_freq when turbo changes
This makes the file behave a little weirdly numberwise if you are changing scaling values.

For example, --min 10 with turbo on could be 460mhz, but with turbo off would be 400mhz which is weird.

Reading from this alternate file means that 100 is always 100% of the reported CPU freq, which for amd_pstate includes boost clocks. As such, if --turbo is off, your --max 100 will be bound to the max non-boost frequency